Introduce environments, which are top-level trees that contain all tokens.
Using an environment as a parent to the parse trees, as opposed to the Node
class, allows the storage of all the tokens in existence to be placed in an
object that can be garbage collected. Every Parser object defines an environment
internally that will span all trees produced by the parser.

Tests reveal this is slightly slower, by 1-2 minutes than the test mode (which
skipped generating the parse tree) on all ~47,000 Nessus plugins. Due to the
small difference, and the fact that memory usage is now constant when operating
on the whole Nessus plugin directory, the test mode has been removed.