Prevent a type error when generating context with some Node types.
Certain Node types (Call, Lvalue, etc.) appear to have nested arrays
of tokens, instead of a single array. Producing context for one of
these Node subclasses would end up causing a crash, when 'region' is
called on an Array. Example:

  tree = Nasl::Parser.new.parse("if(description){exit.foo();}")
  exit_call = tree.all(:Call).first
  if_stmt   = tree.all(:If).first
  puts exit_call.context(if_stmt)

I am uncertain if the nested arrays of tokens are due to a grammar
inconsistency or something else, so I'm not sure if this is the right
place to fix it. That said, it does prevent the crash.