WIP: Allow unquoted keys in arrays in a backwards-compatible way.
Users of this library may expect Arrays to only have Strings (DATA or STRING)
for keys, so we convert unquoted keys (which use the Identifier grammar rule)
to a "fake" String (an OpenStruct that has the same methods as a String).

This is too clever; doing a .context() on one of these "fake" Strings will fail
right now. Getting it to work will require mocking even more of what lexer
does, which is not clean at all.